rearranged place where renaming happens
