+++ library/kani_macros/src/sysroot/contracts/check.rs
@@ -9,9 +9,7 @@ use syn::{Expr, FnArg, ItemFn, Token};
 
 use super::{
     helpers::*,
-    shared::{
-        build_ensures, count_remembers, make_unsafe_argument_copies, try_as_result_assign_mut,
-    },
+    shared::{build_ensures, count_remembers, try_as_result_assign_mut},
     ContractConditionsData, ContractConditionsHandler, INTERNAL_RESULT_IDENT,
 };
 
@@ -36,9 +34,9 @@ impl<'a> ContractConditionsHandler<'a> {
                     #(#inner)*
                 )
             }
-            ContractConditionsData::Ensures { argument_names, attr } => {
-                let (arg_copies, copy_clean) = make_unsafe_argument_copies(&argument_names);
-                let (remembers_stmts, ensures_clause) = build_ensures(attr, remember_count);
+            ContractConditionsData::Ensures { attr } => {
+                let (arg_copies, copy_clean, ensures_clause) =
+                    build_ensures(&self.annotated_fn.sig, attr, remember_count);
 
                 // The code that enforces the postconditions and cleans up the shallow
                 // argument copies (with `mem::forget`).
@@ -56,7 +54,6 @@ impl<'a> ContractConditionsHandler<'a> {
                 let result = Ident::new(INTERNAL_RESULT_IDENT, Span::call_site());
                 quote!(
                     #arg_copies
-                    #remembers_stmts
                     #(#inner)*
                     #exec_postconditions
                     #result
